backend/requirements.txt
flask
flask-cors
langchain
faiss-cpu
openai
